ref. re-design struct
