+++ .gitignore
@@ -1,5 +1,3 @@
 .vscode/
-bulletin.c
-임시저장용.c
-posts.txt
-users.txt
+# *.txt
+main
