--- posts.txt
@@ -1,15 +1,0 @@
-a
-b
-c
-c
-b
-a
-hi
-my name is hyewon
-hyewon
-a
-b
-hyewon
-안녕
-사람이 많다
-임혜원
